Rename Model Files widgets
