+++ moxin-frontend/src/landing/mod.rs
@@ -2,8 +2,8 @@ pub mod download_item;
 pub mod downloads;
 pub mod landing_screen;
 pub mod model_card;
+pub mod model_files;
 pub mod model_files_list;
-pub mod model_files_items;
 pub mod model_files_item;
 pub mod model_files_tags;
 pub mod model_list;
